% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{atrium_darkgray}
\alias{atrium_darkgray}
\title{Atrium Dark Gray}
\format{
An object of class \code{character} of length 1.
}
\usage{
atrium_darkgray
}
\description{
Atrium Dark Gray
}
\keyword{datasets}
