% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{atrium_tealhighlight}
\alias{atrium_tealhighlight}
\title{Atrium Teal Highlight}
\format{
An object of class \code{character} of length 1.
}
\usage{
atrium_tealhighlight
}
\description{
Atrium Teal Highlight
}
\keyword{datasets}
