% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{atrium_tealshadow}
\alias{atrium_tealshadow}
\title{Atrium Teal Shadow}
\format{
An object of class \code{character} of length 1.
}
\usage{
atrium_tealshadow
}
\description{
Atrium Teal Shadow
}
\keyword{datasets}
