% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{bridge_yellowgreen}
\alias{bridge_yellowgreen}
\title{Bridge Yellow-Green}
\format{
An object of class \code{character} of length 1.
}
\usage{
bridge_yellowgreen
}
\description{
Bridge Yellow-Green
}
\keyword{datasets}
