% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fa_lookup}
\alias{fa_lookup}
\title{Font Awesome Lookup}
\format{
\subsection{\code{fa_lookup}}{

A data frame with 2,350 rows and 4 columns:
\describe{
\item{class}{class name (with leading \verb{fa-} removed), all lowercase, no spaces}
\item{name}{Name of icon, title case with spaces}
\item{unicode}{unicode}
\item{group}{group classification of the icon}
}
}
}
\usage{
data(fa_lookup)
}
\description{
Lookup font awesome icons by groups (icons may be duplicated, i.e. belong to
more than one group)
}
\keyword{datasets}
