% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_fontawesome.R
\name{fa_span}
\alias{fa_span}
\title{Create Font Awesome html span}
\usage{
fa_span(icon_class, solid = TRUE)
}
\arguments{
\item{icon_class}{a string or vector of icon class(es) (e.g. head-side-virus, cat, user-doctor)}

\item{solid}{default \code{TRUE} whether to use solid or regular icon (most icons are solid, some have both)}
}
\description{
Takes the font awesome icon class and creates the html span to display the icon
}
