% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{fix_grid}
\alias{fix_grid}
\title{Fix ggplot gridlines}
\usage{
fix_grid(gg, grid = c("none", "x", "y", "both"), facet = FALSE)
}
\arguments{
\item{gg}{a ggplot}

\item{grid}{string, which gridlines to show (x, y, none, both) default 'both'}

\item{facet}{logical, default \code{FALSE} will add axis lines & ticks when grids are dropped (not need for theme_xxx_facet where full panel.border & axis ticks are already present)}
}
\description{
Specify which gridlines to show in ggplot
}
