% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_flex_table.R
\name{format_flex_table}
\alias{format_flex_table}
\title{Format a Flextable}
\usage{
format_flex_table(ft, width_perc = 100, body_font = 10)
}
\arguments{
\item{ft}{a flextable object, the flextable to be formatted.}

\item{width_perc}{a numeric between 1 and 100 indicating the percent width
to expand to fill with a default of \code{100}.}

\item{body_font}{a numeric, the font of the table body with default of \code{10}.}
}
\value{
a formatted flextable
}
\description{
Format a Flextable
}
