% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_percent.R
\name{format_perc}
\alias{format_perc}
\title{Format Percentage to Significant Digits}
\usage{
format_perc(x, digitz = 1)
}
\arguments{
\item{x}{a numeric to be formatted}

\item{digitz}{an integer the number of digits to be formatted with default of \code{1}}
}
\value{
a formatted string
}
\description{
Format Percentage to Significant Digits
}
\examples{

format_perc(.0012343)
# Preserved significant digit
format_perc(.001)
}
