% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_fontawesome.R
\name{geom_fontawesome}
\alias{geom_fontawesome}
\title{Use Font Awesome icons as ggplot geom}
\usage{
geom_fontawesome(icon_class, solid = TRUE, label.color = NA, fill = NA, ...)
}
\arguments{
\item{icon_class}{the icon class (e.g. head-side-virus, cat, user-doctor), can be a single string or a column in your plot data that contains valid icon strings}

\item{solid}{default \code{TRUE} whether to use solid or regular icon (most icons are solid, some have both)}

\item{label.color}{default \code{NA} does not draw a outline around the icon (a typical ggplot2 label style); passed to \code{\link[ggtext]{geom_richtext}}}

\item{fill}{default \code{NA} does not fill in space around the icon (transparent); passed to \code{\link[ggtext]{geom_richtext}}}

\item{...}{other arguments passed to \code{\link[ggtext]{geom_richtext}}}
}
\description{
A ggplot2 geom that displays font awesome icons. Requires libraries
\code{showtext} and \code{ggtext}. See
\href{https://fontawesome.com/v6/icons?o=a&m=free}{Font Awesome Free V6 Icons List}
for a full list of icons available. Brands are currently NOT available. For
more control of the aesthetics, you may need to use \code{\link[ggtext]{geom_richtext}()}
directly and supply the full \code{\link[ggplot2]{aes}()} yourself, i.e.
\code{geom_richtext(aes(label = fa_span(column_name), ...), label.color = NA, fill = NA)}
}
