% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{id_bs_theme}
\alias{id_bs_theme}
\title{Wake Forest Atrium ID Bootstrap Theme}
\usage{
id_bs_theme(font_scale = 0.85)
}
\arguments{
\item{font_scale}{default \code{0.85}, a scalar multiplier to apply to the base font size. For example, a value of 1.5 scales font sizes to 150\% and a value of 0.8 scales to 80\%. Must be a positive number.}
}
\description{
Creates a Bootstrap theme object with WF Atrium ID colors.
To overwrite options or edit other theme options, pass the theme to \code{bslib::bs_theme_update()}.
}
