% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertLogoTextAddin.R
\name{insertLogoTextHTMLAddin}
\alias{insertLogoTextHTMLAddin}
\title{Insert IDEAS Text Logo in HTML}
\usage{
insertLogoTextHTMLAddin()
}
\description{
Call this function as an addin to insert the IDEAS text logo (HTML syntax) at the cursor position.
}
