% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertLogoTextAddin.R
\name{insertLogoTextMDAddin}
\alias{insertLogoTextMDAddin}
\title{Insert IDEAS Text Logo in Markdown.}
\usage{
insertLogoTextMDAddin()
}
\description{
Call this function as an addin to insert the IDEAS text logo (markdown syntax) at the cursor position.
}
