% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{logoTextHTML}
\alias{logoTextHTML}
\title{IDEAS text Logo in HTML format}
\format{
\subsection{\code{logoTextHTML}}{

A character string of length 1
}
}
\usage{
data(logoTextHTML)
}
\description{
IDEAS text Logo as in string in HTML format
}
\keyword{datasets}
