% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{logoTextMD}
\alias{logoTextMD}
\title{IDEAS text Logo in MD format}
\format{
\subsection{\code{logoTextMD}}{

A character string of length 1
}
}
\usage{
data(logoTextMD)
}
\description{
IDEAS text Logo as in string in Markdown format
}
\keyword{datasets}
