% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_package.R
\name{make_package}
\alias{make_package}
\title{Make an R Package}
\usage{
make_package(path = here::here(), shiny = FALSE)
}
\arguments{
\item{path}{Path automatically set by id_project.dcf (see
\code{./rstudio/templates/project/})}

\item{shiny}{default \code{FALSE} does not include a shiny application in the package}
}
\value{
Returns nothing.  See description above.
}
\description{
This function makes an R package for ID with standard folder
structure and option for shiny application.
}
\details{
Behind the scenes, this function used by id_package.dcf when
a user selects New project... > New Directory > ID R Package and Shiny Templates
within the RStudio IDE. See \code{./rstudio/templates/project/}.
}
\examples{
\dontrun{
  make_package(path = "~/test_project")
}
\dontrun{
  make_package(path = "~/test_project", shiny = TRUE))
}
}
