% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_project.R
\name{make_project}
\alias{make_project}
\title{Make an Analysis Project}
\usage{
make_project(
  path = here::here(),
  report = TRUE,
  website = FALSE,
  book = FALSE,
  dashboard = FALSE
)
}
\arguments{
\item{path}{Path automatically set by id_project.dcf (see
\code{./rstudio/templates/project/})}

\item{report}{default: \code{TRUE}, whether to include a \code{.qmd} report for a html, pdf, or docx report}

\item{website}{default: \code{FALSE}, whether to create a quarto website}

\item{book}{default: \code{FALSE}, whether to create a quarto book}

\item{dashboard}{default: \code{FALSE}, whether to create a quarto dashboard}
}
\value{
Returns nothing.  See description above.
}
\description{
This function makes an R project for ID research with standard
folder structure and varying output options.
}
\details{
Behind the scenes, this function used by id_project.dcf when
a user selects New project... > New Directory > ID Research Project Template
within the RStudio IDE. See \code{./rstudio/templates/project/}.
}
\examples{
\dontrun{
  make_project(path = "~/test_project")
}
\dontrun{
  make_project(path = "~/test_project", website = TRUE))
}
}
