% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_nice_date.R
\name{nice_date}
\alias{nice_date}
\title{Format a IS18001 Date to US Convention}
\usage{
nice_date(x)
}
\arguments{
\item{x}{a date object}
}
\value{
a date string
}
\description{
Format a IS18001 Date to US Convention
}
