% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{palette_set_atrium}
\alias{palette_set_atrium}
\title{Set ggplot2 palettes to Atrium defaults}
\usage{
palette_set_atrium(
  discrete.colour = scale_discrete_atrium,
  discrete.fill = scale_discrete_atrium,
  continuous.color = scale_color_continuous_atrium,
  continuous.fill = scale_fill_continuous_atrium
)
}
\arguments{
\item{discrete.colour}{default \code{\link[idstyle]{scale_discrete_atrium}}, sets option ggplot2.discrete.colour}

\item{discrete.fill}{default \code{\link[idstyle]{scale_discrete_atrium}}, sets option ggplot2.discrete.fill}

\item{continuous.color}{default \code{\link[idstyle]{scale_color_continuous_atrium}}, sets option ggplot2.continuous.colour}

\item{continuous.fill}{default \code{\link[idstyle]{scale_fill_continuous_atrium}}, sets option ggplot2.continuous.fill}
}
\description{
The default ggplot2 color/fill scale options for Atrium ID.
}
