% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{palette_set_gg}
\alias{palette_set_gg}
\title{Reset palettes to ggplot2 defaults}
\usage{
palette_set_gg(
  discrete.colour = NULL,
  discrete.fill = NULL,
  continuous.color = NULL,
  continuous.fill = NULL
)
}
\arguments{
\item{discrete.colour}{default \code{NULL}, sets option ggplot2.discrete.colour}

\item{discrete.fill}{default \code{NULL}, sets option ggplot2.discrete.fill}

\item{continuous.color}{default \code{NULL}, sets option ggplot2.continuous.colour}

\item{continuous.fill}{default \code{NULL}, sets option ggplot2.continuous.fill}
}
\description{
The default ggplot2 color/fill scale options
}
