% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{palette_set_id}
\alias{palette_set_id}
\title{Set ggplot2 palettes to ID defaults}
\usage{
palette_set_id(
  discrete.colour = scale_discrete_id,
  discrete.fill = scale_discrete_id,
  continuous.color = scale_color_continuous_id,
  continuous.fill = scale_fill_continuous_id
)
}
\arguments{
\item{discrete.colour}{default \code{\link[idstyle]{scale_discrete_id}}, sets option ggplot2.discrete.colour}

\item{discrete.fill}{default \code{\link[idstyle]{scale_discrete_id}}, sets option ggplot2.discrete.fill}

\item{continuous.color}{default \code{\link[idstyle]{scale_color_continuous_id}}, sets option ggplot2.continuous.colour}

\item{continuous.fill}{default \code{\link[idstyle]{scale_fill_continuous_id}}, sets option ggplot2.continuous.fill}
}
\description{
The default ggplot2 color/fill scale options for ID.
}
