% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{palette_set_wake}
\alias{palette_set_wake}
\title{Set ggplot2 palettes to Wake defaults}
\usage{
palette_set_wake(
  discrete.colour = scale_discrete_wake,
  discrete.fill = scale_discrete_wake,
  continuous.color = scale_color_continuous_wake,
  continuous.fill = scale_fill_continuous_wake
)
}
\arguments{
\item{discrete.colour}{default \code{\link[idstyle]{scale_discrete_wake}}, sets option ggplot2.discrete.colour}

\item{discrete.fill}{default \code{\link[idstyle]{scale_discrete_wake}}, sets option ggplot2.discrete.fill}

\item{continuous.color}{default \code{\link[idstyle]{scale_color_continuous_wake}}, sets option ggplot2.continuous.colour}

\item{continuous.fill}{default \code{\link[idstyle]{scale_fill_continuous_wake}}, sets option ggplot2.continuous.fill}
}
\description{
The default ggplot2 color/fill scale options for Wake ID.
}
