% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_figures.R
\name{pull_figures}
\alias{pull_figures}
\title{Pull Figures}
\usage{
pull_figures(dir = here::here("output"))
}
\arguments{
\item{dir}{directory with figures}
}
\value{
a list of plot items
}
\description{
Will pull figures from a defined folder if they are named "fig_xx"
}
