% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_tables.R
\name{pull_tables}
\alias{pull_tables}
\title{Pull Tables}
\usage{
pull_tables(dir = here::here("output"))
}
\arguments{
\item{dir}{directory with tables}
}
\value{
a list of tables items
}
\description{
Will pull figures from a defined folder if they are named "tab_xx"
}
