% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\name{register_font_awesome}
\alias{register_font_awesome}
\title{Register Font Awesome Fonts for Use}
\usage{
register_font_awesome(verbose = T)
}
\arguments{
\item{verbose}{default \code{TRUE}}
}
\description{
Registers Font Awesome 6 Free with package \code{systemfonts}. Use
Font Awesome regular icons with font-family 'fa' and solid icons
with font-family 'fa-solid'
}
