% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly.R
\name{remove_plotly_dups}
\alias{remove_plotly_dups}
\title{Remove plotly duplicates in legend}
\usage{
remove_plotly_dups(p)
}
\arguments{
\item{p}{a plotly}
}
\description{
Removes duplicates from a plotly legend where groups are shown in parentheses and numbered
}
