% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{scale_color_continuous_id}
\alias{scale_color_continuous_id}
\title{ID Continuous Color Scale Function}
\usage{
scale_color_continuous_id(option = "mako", end = 0.8, ...)
}
\arguments{
\item{option}{default 'mako', a character string indicating the color map option to use.}

\item{end}{The (corrected) hue in \code{⁠[0,1]}⁠ at which the color map ends.
default \code{0.8} so that the whole color map is visible on a white background.}

\item{...}{other arguments passed to \code{\link[ggplot2]{scale_color_gradientn}()}}
}
\description{
Continuous gradient scales using 'mako' viridis palette
}
