% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{scale_color_continuous_wake}
\alias{scale_color_continuous_wake}
\title{Wake Continuous Color Scale Function}
\usage{
scale_color_continuous_wake(...)
}
\arguments{
\item{...}{other arguments passed to \code{\link[ggplot2]{scale_color_gradientn}()}}
}
\description{
Continuous gradient scales of black & Wake gold
}
