% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{scale_discrete_atrium}
\alias{scale_discrete_atrium}
\title{Atrium Discrete Scale Function}
\format{
An object of class \code{list} of length 2.
}
\usage{
scale_discrete_atrium
}
\description{
Atrium Discrete Scale Function
}
\keyword{datasets}
