% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{scale_discrete_wake2}
\alias{scale_discrete_wake2}
\title{Wake Discrete Scale Function}
\format{
An object of class \code{list} of length 1.
}
\usage{
scale_discrete_wake2
}
\description{
Wake Discrete Scale Function
}
\keyword{datasets}
