% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{scale_fill_continuous_atrium}
\alias{scale_fill_continuous_atrium}
\title{Atrium Continuous Fill Scale Function}
\usage{
scale_fill_continuous_atrium(...)
}
\arguments{
\item{...}{other arguments passed to \code{\link[ggplot2]{scale_color_gradientn}()}}
}
\description{
Continuous gradient scales of gray & Atrium teal
}
