% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{scale_fill_continuous_id}
\alias{scale_fill_continuous_id}
\title{ID Continuous Fill Scale Function}
\usage{
scale_fill_continuous_id(option = "mako", ...)
}
\arguments{
\item{option}{default 'mako', a character string indicating the color map option to use.}

\item{...}{other arguments passed to \code{\link[ggplot2]{scale_color_gradientn}()}}
}
\description{
Continuous gradient scales using 'mako' viridis palette
}
