% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{scale_fill_continuous_wake_atrium}
\alias{scale_fill_continuous_wake_atrium}
\title{Wake-Atrium Bridge Continuous Fill Scale}
\usage{
scale_fill_continuous_wake_atrium(...)
}
\arguments{
\item{...}{other arguments passed to \code{\link[ggplot2]{scale_color_gradientn}()}}
}
\description{
Continuous gradient scales that bridges Wake gold & Atrium teal with yellow-green
}
