% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig_round.R
\name{siground}
\alias{siground}
\title{Round to Significant Digits}
\usage{
siground(x, digitz = 1)
}
\arguments{
\item{x}{a numeric to be formatted}

\item{digitz}{an integer representing number of digits to maintain}
}
\description{
Preserves trailing zeros when rounding to significant digits
}
