% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{theme_atrium}
\alias{theme_atrium}
\title{Atrium ID ggplot2 theme}
\usage{
theme_atrium(
  grid = "none",
  base_size = 11,
  base_family = wake_system_fonts$sans_serif,
  base_line_size = base_size/22,
  base_rect_size = base_size/22,
  ...
)
}
\arguments{
\item{grid}{string, which gridlines to show (x, y, none, both) default 'none'}

\item{base_size}{default \code{11}}

\item{base_family}{default \code{idstyle::wake_system_fonts$sans_serif}}

\item{base_line_size}{default \code{base_size/22}}

\item{base_rect_size}{default \code{base_size/22}}

\item{...}{other parameters passed to \code{\link[idstyle]{theme_atrium_elements}} or
\code{\link[ggplot2]{theme}()}}
}
\description{
The default ggplot2 theme for Atrium ID Research
}
