% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{theme_atrium_elements}
\alias{theme_atrium_elements}
\title{Atrium ID ggplot2 theme elements}
\usage{
theme_atrium_elements(
  title_color = atrium_teal,
  title_family = wake_system_fonts$sans_serif,
  title_face = "bold",
  subtitle_family = wake_system_fonts$sans_serif,
  subtitle_face = "plain",
  caption_family = wake_system_fonts$sans_serif,
  caption_face = "plain",
  subtitle_color = atrium_darkgray,
  caption_color = "black",
  legend_title_color = "black",
  legend_title_face = "plain",
  axis_title_color = "black",
  axis_title_face = "plain",
  strip_fill = atrium_teal,
  ...
)
}
\arguments{
\item{title_color}{default \code{idstyle::atrium_teal}}

\item{title_family}{default \code{idstyle::wake_system_fonts$sans_serif}}

\item{title_face}{default 'bold'}

\item{subtitle_family}{default \code{idstyle::wake_system_fonts$sans_serif}}

\item{subtitle_face}{default 'plain'}

\item{caption_family}{default \code{idstyle::wake_system_fonts$sans_serif}}

\item{caption_face}{default 'plain'}

\item{subtitle_color}{default \code{idstyle::atrium_darkgray}}

\item{caption_color}{default 'black'}

\item{legend_title_color}{default 'black'}

\item{legend_title_face}{default 'plain'}

\item{axis_title_color}{default 'black'}

\item{axis_title_face}{default 'plain'}

\item{strip_fill}{default \code{idstyle::atrium_teal}}

\item{...}{other parameters passed to
\code{\link[idstyle]{theme_wake_elements}()} or
\code{\link[ggplot2]{theme}()}}
}
\description{
The ggplot2 theme element for Atrium. Returns \code{link[ggplot2]{theme}()} with lots of defaults for Atrium ID Research
}
