% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{theme_set_atrium}
\alias{theme_set_atrium}
\title{Set Atrium theme & palette for ggplot2}
\usage{
theme_set_atrium()
}
\description{
The default ggplot2 theme (\code{link[idstyle]{theme_atrium}()}) set along
with default color/fill scale options, \code{link[idstyle]{palette_set_atrium}()}
}
