% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{theme_set_id}
\alias{theme_set_id}
\title{Set ID theme & palette for ggplot2}
\usage{
theme_set_id()
}
\description{
The default ggplot2 theme (\code{link[idstyle]{theme_id}()}) set along
with default color/fill scale options, \code{link[idstyle]{palette_set_id}()}
}
