% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{theme_set_wake}
\alias{theme_set_wake}
\title{Set Wake theme & palette for ggplot2}
\usage{
theme_set_wake()
}
\description{
The default ggplot2 theme (\code{link[idstyle]{theme_wake}()}) set along
with default color/fill scale options, \code{link[idstyle]{palette_set_wake}()}
}
