% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{theme_wake_elements}
\alias{theme_wake_elements}
\title{Wake ID ggplot2 theme elements}
\usage{
theme_wake_elements(
  title_size = 18,
  title_color = wake_gold,
  title_family = wake_google_fonts$sans_serif,
  title_face = "plain",
  subtitle_size = 16,
  subtitle_color = wake_gray,
  subtitle_family = wake_google_fonts$serif,
  subtitle_face = "bold",
  caption_size = 9,
  caption_color = wake_gray,
  caption_family = wake_google_fonts$serif,
  caption_face = "bold",
  legend_title_size = 11,
  legend_title_color = wake_gray,
  legend_title_face = "bold",
  axis_title_size = 12,
  axis_title_color = wake_gray,
  axis_title_face = "bold",
  legend_size = 10,
  strip_fill = "black",
  strip_color = "white",
  strip_size = 13,
  strip_face = "bold",
  ...
)
}
\arguments{
\item{title_size}{default \code{18}}

\item{title_color}{default \code{idstyle::wake_gold}}

\item{title_family}{default \code{idstyle::wake_google_fonts$serif}}

\item{title_face}{default 'bold'}

\item{subtitle_size}{default \code{16}}

\item{subtitle_color}{default \code{idstyle::wake_gray}}

\item{subtitle_family}{default \code{title_family}}

\item{subtitle_face}{default 'bold'}

\item{caption_size}{default \code{9}}

\item{caption_color}{default \code{idstyle::wake_gray}}

\item{caption_family}{default \code{title_family}}

\item{caption_face}{default 'bold'}

\item{legend_title_size}{default \code{11}}

\item{legend_title_color}{default \code{idstyle::wake_gray}}

\item{legend_title_face}{default 'bold'}

\item{axis_title_size}{default \code{13}}

\item{axis_title_color}{default \code{idstyle::wake_gray}}

\item{axis_title_face}{default 'bold'}

\item{legend_size}{default \code{10}}

\item{strip_fill}{default 'black'}

\item{strip_color}{default 'white'}

\item{strip_size}{default \code{13}}

\item{strip_face}{default 'bold'}

\item{...}{other parameters passed to \code{\link[ggplot2]{theme}()}}
}
\description{
The ggplot2 theme element for wake. Returns \code{link[ggplot2]{theme}()} with lots of defaults for Wake ID Research
}
