% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{update_geoms_gg}
\alias{update_geoms_gg}
\title{Update ggplot2 geom defaults}
\usage{
update_geoms_gg(line_color = "black", point_color = "black", point_size = 1.5)
}
\arguments{
\item{line_color}{default 'black'}

\item{point_color}{default 'black'}

\item{point_size}{default \code{1.5}}
}
\description{
Update ggplot2 geom defaults; the default here are ggplot defaults
}
