% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useWakeCss.R
\name{useWakeCss}
\alias{useWakeCss}
\title{Use Wake CSS}
\usage{
useWakeCss(x)
}
\arguments{
\item{x}{html widget object}
}
\description{
Applies header info via \code{link[htmlwidgets]{prependContent}()}
for using wake css within any html widget (e.g. leaflet map, plotly graph).
\strong{This is only needed for stand alone widgets}, as widgets within an html page
(such as quarto \code{idreport}) will have this css applied to the document already.
}
