% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{wake_all_colors}
\alias{wake_all_colors}
\title{Wake Forest Atrium Full Color Guide}
\format{
An object of class \code{list} of length 28.
}
\usage{
wake_all_colors
}
\description{
Wake Forest Atrium Full Color Guide
}
\keyword{datasets}
