% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{wake_cols}
\alias{wake_cols}
\title{Wake Forest University Color Guide}
\format{
An object of class \code{character} of length 2.
}
\usage{
wake_cols
}
\description{
Wake Forest University Color Guide
}
\keyword{datasets}
