% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{wake_darkgold}
\alias{wake_darkgold}
\title{Wake Forest Dark Gold}
\format{
An object of class \code{character} of length 1.
}
\usage{
wake_darkgold
}
\description{
Wake Forest Dark Gold
}
\keyword{datasets}
