% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\docType{data}
\name{wake_font_collections}
\alias{wake_font_collections}
\title{Wake Forest Atrium Font Collections}
\format{
An object of class \code{list} of length 2.
}
\usage{
wake_font_collections
}
\description{
contains a list of font collections via \code{bslib::font_collection()}
}
\keyword{datasets}
