% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\docType{data}
\name{wake_google_fonts}
\alias{wake_google_fonts}
\title{Wake Forest Atrium Google Fonts}
\format{
An object of class \code{list} of length 2.
}
\usage{
wake_google_fonts
}
\description{
contains a list of character strings of Google fonts
}
\keyword{datasets}
