% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{wake_oldgold}
\alias{wake_oldgold}
\title{Wake Forest Old Gold}
\format{
An object of class \code{character} of length 1.
}
\usage{
wake_oldgold
}
\description{
Wake Forest Old Gold
}
\keyword{datasets}
