% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\docType{data}
\name{wake_sansserif}
\alias{wake_sansserif}
\title{Wake Forest Serif Font}
\format{
An object of class \code{character} of length 1.
}
\usage{
wake_sansserif
}
\description{
a character string of the sans serif font families, comma delimited
}
\keyword{datasets}
