% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{wake_secondary}
\alias{wake_secondary}
\title{Wake Forest Secondary Colors}
\format{
An object of class \code{character} of length 7.
}
\usage{
wake_secondary
}
\description{
Wake Forest Secondary Colors
}
\keyword{datasets}
