% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\docType{data}
\name{wake_system_fonts}
\alias{wake_system_fonts}
\title{Wake Forest Atrium System Fonts}
\format{
An object of class \code{list} of length 2.
}
\usage{
wake_system_fonts
}
\description{
contains a list of character strings of system fonts
}
\keyword{datasets}
