% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{wake_yellow}
\alias{wake_yellow}
\title{Wake Forest Yellow}
\format{
An object of class \code{character} of length 1.
}
\usage{
wake_yellow
}
\description{
Wake Forest Yellow
}
\keyword{datasets}
