% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_flex_table.R
\name{format_flex_table}
\alias{format_flex_table}
\title{Format a Flextable}
\usage{
format_flex_table(
  ft,
  width_perc = 80,
  body_font = 10,
  header_font = 10,
  color = c("none", "id", "wake", "atrium")
)
}
\arguments{
\item{ft}{a flextable object, or an object that can be coerced into a flextable
object, to be formatted.}

\item{width_perc}{a numeric between 1 and 100 indicating the percent width
to expand to fill with a default of \code{80}. \code{NULL} for automatic
flextable sizing.}

\item{body_font}{a numeric, the font of the table body with default of \code{10}.}

\item{header_font}{a numeric, the font of the table header with default of \code{10}.}

\item{color}{a string, default 'none' returns a black and white table,
for 'wake' and 'atrium' a gray striped table with a gold or teal
background header}
}
\value{
a formatted flextable
}
\description{
Format a Flextable
}
